page for adding emergency contacts. Static at this point, but includes form validation
includes new browserRouter + switch in App

Refactoring and enhancements. Don't enable "add new phone number" unless a valid number is already present in the form.

Reasoning: This prevents the form from growing without cause
Refactor: A sub-component is added to simplify the way the code reads in the add contact numbers form
Refactor: A helper component is added for error display to further DRY the code