Add action button for "Archive Properties"
By default, the button is disabled and it only becomes active when one or more properties have their checkboxes selected.
After the archive request is complete, reset the table element and update the list on the page
New state tracks the selected properties.
archivedCallCount is a workaround to force a re-render on the bootstrap table
css cleanup
selectRow object needs to have access to state, so it moves within the scope of the component.