Gitignore `.env` file (#679)
gitignoring the .env file. This file should be ignored as devs can have
different env vars, and they should be kept secret.

Created a `.env.example` file for people to copy to a `.env` file to get
them started.

Updated the Install docs to reflect this.

Added the option to run the backend off of a remote web server.