Remove race condition in async functions when updating state
The `handleConfirmButton` awaits on the `updateProperty` function to
update the property, and then does an additional call to get the
property. The issue is that the await keyword doesn't really impact
anything here, because the `updateProperty` function is also an async
function. So what was happening was that sometimes the property would be
updated with stale information due to the `updateProperty` and
`getProperty` racing to fetch data and update the state.

Removed the `getProperty` call as it is not needed. The `updateProperty`
method already updates the state.

Fixes #677