Refactor dashboard component
Refactored dashboard to work with the new dashboard api in the backend.
A noticeable performance boost since the frontend client no longer
iterates through all tenants in the system, when the backend db is
perfectly capable of outputting the required data.

Fixes #730 Display staff assignments
Fixes #731 Display at most 5 staff assignments so that the save button
is visible.

There was an issue with saving staff to a tenant. This was caused by
checking if the tenant attribute staff was assigned, however that
attribute will be an empty array when staff is not assigned which is a
truthy value in JS. Changed it to `staffID`.

Updated the OpenTickets and New Managers widgets component so that they
columns will always be aligned no matter the size of the stat column.
It now uses CSS Grid to accomplish this.

Created test for assigning staff to tenant via the dashboard page.
Created test for checking staff shows up in unhoused section of the
tenants view page. This uncovered a bug which was patched. The code was
assuming the tenant has a lease, but unhoused tenants do not have
leases.