Combine the frontend and backend into a single app
This will improve security since we can use a session and an encrypted
cookie to store the user credentials for accessing the app.

This is done by serving the React frontend and api calls from a single
server.

The backend has been rebuilt with Rails, which will allow us to take
advantage of its built in security features, and leverage built-in tools
that will ease developing efforts, and keep maintenance costs low.